+++ src/main/java/org/sourcelab/kafka/connect/apiclient/request/put/PutConnectorPluginConfigValidate.java
@@ -17,7 +17,6 @@
 
 package org.sourcelab.kafka.connect.apiclient.request.put;
 
-import com.google.common.base.Preconditions;
 import org.sourcelab.kafka.connect.apiclient.request.JacksonFactory;
 import org.sourcelab.kafka.connect.apiclient.request.dto.ConnectorPluginConfigValidationResults;
 
@@ -25,6 +24,7 @@ import java.io.IOException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 
 import static com.google.common.net.UrlEscapers.urlPathSegmentEscaper;
 
@@ -41,8 +41,8 @@ public final class PutConnectorPluginConfigValidate implements PutRequest<Connec
      * @param config Configuration entries to validate.
      */
     public PutConnectorPluginConfigValidate(final String connectorPluginName, final Map<String, String> config) {
-        Preconditions.checkNotNull(connectorPluginName);
-        Preconditions.checkNotNull(config);
+        Objects.requireNonNull(connectorPluginName);
+        Objects.requireNonNull(config);
         this.connectorPluginName = connectorPluginName;
         this.config = Collections.unmodifiableMap(new HashMap<>(config));
     }