+++ CHANGELOG.md
@@ -3,7 +3,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.0.4 (UNRELEASED)
+
+### Internal Dependency Updates
 - Updated Guava from 24.0-JRE to 25.0-JRE for [CVE-2018-10237](https://github.com/google/guava/wiki/CVE-2018-10237).
+- Replace internal calls to Guava's Preconditions with Objects.requireNonNull()
 
 ## 1.0.3 (04/16/2018)
 - More specific exception classes thrown when requests fail.